Refactor el módulo de verificación de URLs: actualiza las importaciones y agrega la lógica de verificación en un nuevo archivo de servicio.
